+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/) 
 
 ## [UNRELEASED]
 
+### Added
+
+- Added support for regular expressions to `/pattern check`.
+
 ### Mods Added
 
 - HexThings: 0.1.4+1.20.1